% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{dist_between_centroids}
\alias{dist_between_centroids}
\title{Compute the distance between group centroids}
\usage{
dist_between_centroids(d, idx1, idx2)
}
\arguments{
\item{d}{A distance matrix object of class `"dist"`.}

\item{idx1}{A vector of items in group 1}

\item{idx2}{A vector of items in group 2}
}
\value{
The distance between group centroids (see details).

It is possible to infer the distance between group centroids directly from
the distances between items in each group.  The `adonis` test in the ecology
package `vegan` takes advantage of this approach to carry out an ANOVA-like
test on distances.

The approach rests on the assumption that the items in the distance matrix
occupy some high-dimensional Euclidean space.  However, we do not have to
actually create the space to find the distance between centroids.  Using the
assumption that such a space exists, we can use an algebraic formula to find
the centroid distance.

The formulas for this were presented by Apostol and Mnatsakanian in 2003,
though we need to re-arrange equation 28 to get the value we want:

\deqn{| c_1 - c_2 | = \sqrt{
   \frac{1}{n_1 n_2} \sum_{(1,2)} -
   \frac{1}{n_1^2} \sum_{(1)} -
   \frac{1}{n_2^2} \sum_{(2)}},}

where \eqn{n_1} is the number of samples in group 1, \eqn{\sum_{(1)}} is the
sum of squared distances between items in group 1, and \eqn{\sum_{(1,2)}} is
the sum of squared distances between items in group 1 and those in group 2.
}
\description{
Compute the distance between group centroids
}
\references{
Apostol, T.M. and Mnatsakanian, M.A. Sums of squares of distances
  in m-space. Math. Assoc. Am. Monthly 110, 516 (2003).
}
