% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{dist_collapse_centroids}
\alias{dist_collapse_centroids}
\title{Make a new distance matrix of centroid distances between multiple groups}
\usage{
dist_collapse_centroids(d, g, squared = FALSE)
}
\arguments{
\item{d}{A distance matrix object of class \code{dist}.}

\item{g}{A factor representing the groups of items in \code{d}.}

\item{squared}{If \code{TRUE}, return the squared distance between centroids.}
}
\value{
A distance matrix of distances between the group centroids.
}
\description{
Make a new distance matrix of centroid distances between multiple groups
}
