% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_format.R
\name{dist_long}
\alias{dist_long}
\title{Compute distances on data in long format}
\usage{
dist_long(data, obs_col, feature_col, value_col, distance_fcn)
}
\arguments{
\item{data}{A data frame with numerical values in long format.}

\item{obs_col}{The column listing the observation, or row of the matrix.}

\item{feature_col}{The column listing the feature, or column of the matrix.}

\item{value_col}{The column listing the value, to be placed inside the
matrix.}

\item{distance_fcn}{The distance function to use. This function should take
  two numeric vectors and return a number representing the distance or
  dissimilarity.

#' This function requires the packages \code{dplyr}, \code{tibble}, and
\code{tidyr} to be installed. If they are not installed, the function will
generate an error, with a message to install the appropriate packages.}
}
\description{
Compute distances on data in long format
}
\examples{
longdata <- expand.grid(
  SampleID = c("A", "B", "C"),
  FeatureID = c("x", "y", "z"))
longdata$Value = sin(1:9)
longdata

manhattan_distance <- function (v1, v2) sum(abs(v1 - v2))
dist_long(longdata, SampleID, FeatureID, Value, manhattan_distance)
}
