% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{dist_read}
\alias{dist_read}
\title{Read a distance matrix file.}
\usage{
dist_read(file, ...)
}
\arguments{
\item{file}{The file path of the distance matrix to be read}

\item{...}{Additional arguments passed to \code{read.table}}
}
\value{
A dist object
}
\description{
Read a distance matrix file.
}
\details{
Values in the dist object are taken from the lower triangle of the matrix.
}
