% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_setNames}
\alias{dist_setNames}
\title{Set the names/labels of a \code{dist} object.}
\usage{
dist_setNames(d, nm)
}
\arguments{
\item{d}{A distance matrix object of class \code{dist}.}

\item{nm}{New labels for the rows/columns.}
}
\description{
Set the names/labels of a \code{dist} object.
}
\examples{
m4 <- matrix(1:16, nrow=4, dimnames=list(LETTERS[1:4]))
dm4 <- dist(m4)
dist_setNames(dm4, LETTERS[9:12])
}
