% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{dist_write}
\alias{dist_write}
\title{Write a distance matrix file}
\usage{
dist_write(d, file)
}
\arguments{
\item{d}{A dist object}

\item{file}{The file path to write}
}
\value{
Returns the dist object, invisibly
}
\description{
Write a distance matrix file
}
