% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_format.R
\name{pivot_to_matrix}
\alias{pivot_to_matrix}
\alias{pivot_to_numeric_matrix}
\title{Convert a data frame in long format to a matrix}
\usage{
pivot_to_matrix(data, rows_from, cols_from, values_from, fill = 0)

pivot_to_numeric_matrix(data, obs_col, feature_col, value_col)
}
\arguments{
\item{data}{A data frame in long format.}

\item{rows_from}{The column indicating the row of the matrix.}

\item{cols_from}{The column indicating the column of the matrix.}

\item{values_from}{The column indicating the value to be placed inside the
matrix.}

\item{fill}{The value to use for missing combinations of rows and columns.}

\item{obs_col, feature_col, value_col}{The same as \code{rows_from},
\code{cols_from}, and \code{values_from}, respectively.}
}
\description{
Convert a data frame in long format to a matrix
}
\details{
The parameters \code{rows_from}, \code{cols_from}, and \code{values_from}
should be provided as bare column names.

This function requires the packages \code{tidyr}, \code{rlang}, and
\code{tibble} to be installed. If they are not installed, the function will
generate an error, with a message to install the appropriate packages.
}
\section{Functions}{
\itemize{
\item \code{pivot_to_numeric_matrix}: Specialized version for numeric values.
Deprecated; use \code{pivot_to_matrix} instead.
}}

\examples{
longdata <- data.frame(
  sample_id = paste0("Sample", c(1, 1, 1, 2, 2, 3, 3)),
  feature_id = paste0("Feature", c(1, 2, 3, 1, 2, 2, 3)),
  counts = c(132, 41, 7, 56, 11, 929, 83))
pivot_to_matrix(longdata, sample_id, feature_id, counts)
}
