% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\docType{package}
\name{usedist}
\alias{usedist}
\alias{usedist-package}
\title{usedist: a package for working with distance matrices in R}
\description{
In usedist, we provide a number of functions to help with distance matrix
objects, such as those produced by the `dist` function.  Some functions are
geared towards making or altering distance matrix objects.  Others relate to
groups of items in the distance matrix. They provide access to within- or
between-group distances, or use these distances to infer the distance to
group centroids.
}
